requirements.txt
# This file may be used to create an environment using:
# $ conda create --name HH-VAEM --file requirements.txt
# platform: linux-64
pytorch
torchvision
pytorch-lightning
matplotlib
scipy
h5py
scikit-learn
pandas
openpyxl
termcolor
xlrd
seaborn